Creating Common Chart Component to reduce duplication (#218)
