+++ packages/website/package.json
@@ -23,18 +23,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
-    "@types/classnames": "^2.2.10",
-    "@types/jest": "^24.0.0",
-    "@types/lodash": "^4.14.149",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-leaflet": "^2.5.1",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/react-slick": "^0.23.4",
-    "@types/react-swipeable-views": "^0.13.0",
-    "@types/redux-mock-store": "^1.0.2",
     "@typescript-eslint/eslint-plugin": "^2.23.0",
     "@typescript-eslint/parser": "^2.25.0",
     "axios": "^0.19.2",
@@ -94,9 +82,20 @@
     "node": "12.x"
   },
   "devDependencies": {
-    "@types/classnames": "^2.2.10",
     "@types/immutable": "^3.8.7",
     "@types/validator": "^13.1.0",
+    "@types/classnames": "^2.2.10",
+    "@types/jest": "^24.0.0",
+    "@types/lodash": "^4.14.149",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.9.0",
+    "@types/react-dom": "^16.9.0",
+    "@types/react-leaflet": "^2.5.1",
+    "@types/react-redux": "^7.1.7",
+    "@types/react-router-dom": "^5.1.3",
+    "@types/react-slick": "^0.23.4",
+    "@types/react-swipeable-views": "^0.13.0",
+    "@types/redux-mock-store": "^1.0.2",
     "canvas": "^2.6.1"
   }
 }
