+++ packages/website/src/routes/Homepage/ReefTable/SelectedReefCard/index.tsx
@@ -1,178 +1,173 @@
 import React from "react";
 import {
-  Typography,
-  Grid,
-  Button,
-  CardMedia,
-  withStyles,
-  WithStyles,
-  createStyles,
-  Theme,
-  CircularProgress,
-  Card,
-  Hidden,
   Box,
+  Button,
+  Card,
+  CardMedia,
+  CircularProgress,
+  Grid,
+  Hidden,
+  Typography,
 } from "@material-ui/core";
 import { Link } from "react-router-dom";
 
 import { useSelector } from "react-redux";
-import CardChart from "./cardChart";
+import { makeStyles } from "@material-ui/core/styles";
+import { sortByDate } from "../../../../helpers/sortDailyData";
 import { formatNumber } from "../../../../helpers/numberUtils";
 
 import reefImage from "../../../../assets/reef-image.jpg";
 import { degreeHeatingWeeksCalculator } from "../../../../helpers/degreeHeatingWeeks";
 import { reefDetailsSelector } from "../../../../store/Reefs/selectedReefSlice";
 import { Reef } from "../../../../store/Reefs/types";
+import Chart from "../../../../common/Chart";
 
-const styles = (theme: Theme) =>
-  createStyles({
-    card: {
-      padding: 40,
-      [theme.breakpoints.down("xs")]: {
-        padding: 10,
-      },
+const useStyles = makeStyles((theme) => ({
+  card: {
+    padding: 40,
+    [theme.breakpoints.down("xs")]: {
+      padding: 10,
     },
-    cardImage: {
-      borderRadius: "4px 0 0 4px",
-      height: "100%",
+  },
+  cardImage: {
+    borderRadius: "4px 0 0 4px",
+    height: "100%",
 
-      [theme.breakpoints.down("xs")]: {
-        height: 300,
-      },
+    [theme.breakpoints.down("xs")]: {
+      height: 300,
     },
-    metricsContainer: {
-      display: "flex",
-      justifyContent: "space-around",
-      alignItems: "center",
-      flexDirection: "row",
-      margin: theme.spacing(2),
-      width: "100%",
+  },
+  metricsContainer: {
+    display: "flex",
+    justifyContent: "space-around",
+    alignItems: "center",
+    flexDirection: "row",
+    margin: theme.spacing(2),
+    width: "100%",
 
-      [theme.breakpoints.up("lg")]: {
-        flexDirection: "column",
-        alignItems: "start",
-      },
+    [theme.breakpoints.up("lg")]: {
+      flexDirection: "column",
+      alignItems: "start",
     },
-  });
+  },
+}));
 
-type SelectedReefContentProps = WithStyles<typeof styles> & { reef: Reef };
+type SelectedReefContentProps = { reef: Reef };
 
-const SelectedReefContent = withStyles(styles)(
-  ({ classes, reef }: SelectedReefContentProps) => {
-    const {
-      bottomTemperature,
-      degreeHeatingDays,
-      surfaceTemperature,
-      satelliteTemperature,
-    } = reef.liveData;
+const SelectedReefContent = ({ reef }: SelectedReefContentProps) => {
+  const classes = useStyles();
 
-    const surfTemp = surfaceTemperature?.value || satelliteTemperature?.value;
+  const sortedDailyData = sortByDate(reef.dailyData, "date");
+  const dailyDataLen = sortedDailyData.length;
+  const {
+    maxBottomTemperature,
+    surfaceTemperature,
+    satelliteTemperature,
+    degreeHeatingDays,
+  } = sortedDailyData[dailyDataLen - 1];
 
-    const metrics = [
-      {
-        label: "SURFACE TEMP",
-        value: formatNumber(surfTemp, 1),
-        unit: " °C",
-      },
-      {
-        label: `TEMP AT ${reef.depth}m`,
-        value: formatNumber(bottomTemperature?.value, 1),
-        unit: " °C",
-      },
-      {
-        label: "HEAT STRESS",
-        value: formatNumber(
-          degreeHeatingWeeksCalculator(degreeHeatingDays?.value),
-          1
-        ),
-        unit: " DHW",
-      },
-    ];
+  const surfTemp = surfaceTemperature || satelliteTemperature;
 
-    return (
-      <Grid container spacing={1}>
-        <Grid item xs={12} sm={4} lg={3}>
-          <Box position="relative" height="100%">
-            <CardMedia className={classes.cardImage} image={reefImage} />
-            <Hidden smUp>
-              <Box position="absolute" top={16} left={16}>
-                <Typography variant="h5">{reef.name}</Typography>
+  const metrics = [
+    {
+      label: "SURFACE TEMP",
+      value: formatNumber(surfTemp, 1),
+      unit: " °C",
+    },
+    {
+      label: `TEMP AT ${reef.depth}m`,
+      value: formatNumber(maxBottomTemperature, 1),
+      unit: " °C",
+    },
+    {
+      label: "HEAT STRESS",
+      value: formatNumber(degreeHeatingWeeksCalculator(degreeHeatingDays), 1),
+      unit: " DHW",
+    },
+  ];
 
-                {reef.region?.name && (
-                  <Typography variant="h6" style={{ fontWeight: 400 }}>
-                    {reef.region.name}
-                  </Typography>
-                )}
-              </Box>
-            </Hidden>
+  return (
+    <Grid container spacing={1}>
+      <Grid item xs={12} sm={4} lg={3}>
+        <Box position="relative" height="100%">
+          <CardMedia className={classes.cardImage} image={reefImage} />
+          <Hidden smUp>
+            <Box position="absolute" top={16} left={16}>
+              <Typography variant="h5">{reef.name}</Typography>
 
-            <Box position="absolute" bottom={16} right={16}>
-              <Link
-                style={{ color: "inherit", textDecoration: "none" }}
-                to={`/reefs/${reef.id}`}
-              >
-                <Button size="small" variant="contained" color="primary">
-                  EXPLORE
-                </Button>
-              </Link>
+              {reef.region?.name && (
+                <Typography variant="h6" style={{ fontWeight: 400 }}>
+                  {reef.region.name}
+                </Typography>
+              )}
             </Box>
-          </Box>
-        </Grid>
+          </Hidden>
 
-        <Grid
-          item
-          xs={12}
-          sm={8}
-          lg={6}
-          style={{ marginBottom: "1rem", maxHeight: "14rem" }}
-        >
-          <Box pb="0.5rem" pl="0.5rem" fontWeight={400}>
-            <Typography color="textSecondary" variant="subtitle1">
-              MEAN DAILY SURFACE TEMP. (°C)
-            </Typography>
+          <Box position="absolute" bottom={16} right={16}>
+            <Link
+              style={{ color: "inherit", textDecoration: "none" }}
+              to={`/reefs/${reef.id}`}
+            >
+              <Button size="small" variant="contained" color="primary">
+                EXPLORE
+              </Button>
+            </Link>
           </Box>
-          <CardChart
-            dailyData={reef.dailyData}
-            temperatureThreshold={
-              reef.maxMonthlyMean ? reef.maxMonthlyMean + 1 : null
-            }
-            maxMonthlyMean={reef.maxMonthlyMean || null}
-          />
-        </Grid>
-
-        <Grid
-          item
-          xs={12}
-          lg={3}
-          style={{
-            display: "flex",
-          }}
-        >
-          <div className={classes.metricsContainer}>
-            {metrics.map(({ label, value, unit }) => (
-              <div key={label}>
-                <Typography variant="caption" color="textSecondary">
-                  {label}
-                </Typography>
-                <Typography variant="h4" color="primary">
-                  {value}
-                  &nbsp;
-                  <Typography variant="h6" component="span">
-                    {unit}
-                  </Typography>
-                </Typography>
-              </div>
-            ))}
-          </div>
-        </Grid>
+        </Box>
       </Grid>
-    );
-  }
-);
 
-type SelectedReefProps = WithStyles<typeof styles>;
+      <Grid
+        item
+        xs={12}
+        sm={8}
+        lg={6}
+        style={{ marginBottom: "2rem", maxHeight: "14rem" }}
+      >
+        <Box pb="0.5rem" pl="0.5rem" fontWeight={400}>
+          <Typography color="textSecondary" variant="subtitle1">
+            MEAN DAILY SURFACE TEMP. (°C)
+          </Typography>
+        </Box>
+        <Chart
+          dailyData={reef.dailyData}
+          temperatureThreshold={
+            reef.maxMonthlyMean ? reef.maxMonthlyMean + 1 : null
+          }
+          maxMonthlyMean={reef.maxMonthlyMean || null}
+        />
+      </Grid>
 
-const SelectedReefCard = ({ classes }: SelectedReefProps) => {
+      <Grid
+        item
+        xs={12}
+        lg={3}
+        style={{
+          display: "flex",
+        }}
+      >
+        <div className={classes.metricsContainer}>
+          {metrics.map(({ label, value, unit }) => (
+            <div key={label}>
+              <Typography variant="caption" color="textSecondary">
+                {label}
+              </Typography>
+              <Typography variant="h4" color="primary">
+                {value}
+                &nbsp;
+                <Typography variant="h6" component="span">
+                  {unit}
+                </Typography>
+              </Typography>
+            </div>
+          ))}
+        </div>
+      </Grid>
+    </Grid>
+  );
+};
+
+const SelectedReefCard = () => {
+  const classes = useStyles();
   const reef = useSelector(reefDetailsSelector);
 
   const featuredReefId = process.env.REACT_APP_FEATURED_REEF_ID || "";
@@ -201,4 +196,4 @@ const SelectedReefCard = ({ classes }: SelectedReefProps) => {
   ) : null;
 };
 
-export default withStyles(styles)(SelectedReefCard);
+export default SelectedReefCard;
